[dv/top_earlgrey] added regression for chip_sw_lc_ctrl_scrap test
