+++ hw/top_earlgrey/dv/chip_sim_cfg.hjson
@@ -688,12 +688,41 @@
       en_run_modes: ["sw_test_mode_test_rom"]
       reseed: 15
     }
+
     {
-      name: chip_sw_lc_ctrl_raw_to_scrap
-      uvm_test_seq: chip_sw_lc_ctrl_raw_to_scrap_vseq
+      name: chip_sw_lc_ctrl_rma_to_scrap
+      uvm_test_seq: chip_sw_lc_ctrl_scrap_vseq
       sw_images: ["//sw/device/tests/sim_dv:chip_sw_lc_ctrl_scrap_test:1"]
       en_run_modes: ["sw_test_mode_test_rom"]
+      run_opts: ["+bypass_alert_ready_to_end_check=1",
+                 "+src_dec_state=DecLcStRma"]
       reseed: 1
+    }
+    {
+      name: chip_sw_lc_ctrl_raw_to_scrap
+      uvm_test_seq: chip_sw_lc_ctrl_scrap_vseq
+      sw_images: ["//sw/device/tests/sim_dv:chip_sw_lc_ctrl_scrap_test:1"]
+      en_run_modes: ["sw_test_mode_test_rom"]
+      run_opts: ["+bypass_alert_ready_to_end_check=1",
+                 "+src_dec_state=DecLcStRaw"]
+      reseed: 1
+    }
+    {
+      name: chip_sw_lc_ctrl_test_locked0_to_scrap
+      uvm_test_seq: chip_sw_lc_ctrl_scrap_vseq
+      sw_images: ["//sw/device/tests/sim_dv:chip_sw_lc_ctrl_scrap_test:1"]
+      en_run_modes: ["sw_test_mode_test_rom"]
+      run_opts: ["+bypass_alert_ready_to_end_check=1",
+                 "+src_dec_state=DecLcStTestLocked0"]
+      reseed: 1
+    }
+    {
+      name: chip_sw_lc_ctrl_rand_to_scrap
+      uvm_test_seq: chip_sw_lc_ctrl_scrap_vseq
+      sw_images: ["//sw/device/tests/sim_dv:chip_sw_lc_ctrl_scrap_test:1"]
+      en_run_modes: ["sw_test_mode_test_rom"]
+      run_opts: ["+bypass_alert_ready_to_end_check=1"]
+      reseed: 3
     }
     {
       name: chip_sw_lc_walkthrough_dev
@@ -1658,6 +1687,14 @@
 
   // List of regressions.
   regressions: [
+    {
+      name: scrap
+      tests: ["chip_sw_lc_ctrl_raw_to_scrap",
+              "chip_sw_lc_ctrl_rma_to_scrap",
+              "chip_sw_lc_ctrl_test_locked0_to_scrap",
+              "chip_sw_lc_ctrl_rand_to_scrap"
+             ]
+    }
     {
       name: smoke
       tests: ["xbar_chip_smoke",
