+++ sw/device/tests/sim_dv/BUILD
@@ -331,11 +331,10 @@ opentitan_functest(
     ],
 )
 
-cc_library(
-    name = "chip_sw_lc_ctrl_scrap_impl",
-    srcs = ["chip_sw_lc_ctrl_scrap_impl.c"],
-    hdrs = ["chip_sw_lc_ctrl_scrap_impl.h"],
-    target_compatible_with = [OPENTITAN_CPU],
+opentitan_functest(
+    name = "chip_sw_lc_ctrl_scrap_test",
+    srcs = ["chip_sw_lc_ctrl_scrap_test.c"],
+    targets = ["dv"],
     deps = [
         "//hw/top_earlgrey/sw/autogen:top_earlgrey",
         "//sw/device/lib/base:bitfield",
@@ -345,18 +344,10 @@ cc_library(
         "//sw/device/lib/runtime:log",
         "//sw/device/lib/testing:lc_ctrl_testutils",
         "//sw/device/lib/testing/test_framework:check",
-    ],
-)
-
-opentitan_functest(
-    name = "chip_sw_lc_ctrl_scrap_test",
-    srcs = ["chip_sw_lc_ctrl_scrap_test.c"],
-    targets = ["dv"],
-    deps = [
-        ":chip_sw_lc_ctrl_scrap_impl",
         "//sw/device/lib/testing/test_framework:ottf_main",
     ],
 )
+
 cc_library(
     name = "lc_ctrl_transition_impl",
     srcs = ["lc_ctrl_transition_impl.c"],
