--- sw/device/tests/sim_dv/chip_sw_lc_ctrl_scrap_impl.h
@@ -1,12 +1,0 @@
-// Copyright Nuvoton contributors.
-// Licensed under the Apache License, Version 2.0, see LICENSE for details.
-// SPDX-License-Identifier: Apache-2.0
-
-#ifndef OPENTITAN_SW_DEVICE_TESTS_SIM_DV_LC_CTRL_SCRAP_IMPL_H_
-#define OPENTITAN_SW_DEVICE_TESTS_SIM_DV_LC_CTRL_SCRAP_IMPL_H_
-
-#include <stdbool.h>
-
-bool execute_lc_ctrl_scrap_test(bool use_ext_clk);
-
-#endif  // OPENTITAN_SW_DEVICE_TESTS_SIM_DV_LC_CTRL_SCRAP_IMPL_H_
